+++ core/utils/cache.js
@@ -1,6 +1,6 @@
 /*
  * The ti-engine is an open source, free to use—both for personal and commercial projects—framework for the creation of microservice-based solutions using node.js.
- * Copyright © 2021-2023 Boris Kostadinov <kostadinov.boris@gmail.com>
+ * Copyright © 2021-2025 Boris Kostadinov <kostadinov.boris@gmail.com>
  * This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of the License, or (at your option) any later version.
  * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for more details.
  * You should have received a copy of the GNU General Public License along with this program. If not, see <https://www.gnu.org/licenses/>.
@@ -123,7 +123,7 @@ class CommonMemoryCache extends ConnectionObserver {
     }
 
     /**
-     * Used to search for keys by given pattern.
+     * Used to search for keys by a given pattern.
      *
      * @method
      * @param {string} pattern
@@ -379,7 +379,7 @@ class CommonMemoryCache extends ConnectionObserver {
     /**
      * Used to add multiple values to multiple sets in one transactional request.
      * <br/>
-     * NOTE: The two arrays of keys and values must have correct index relations (i.e. first pair on keys[0] and values[0] and so on)!
+     * NOTE: The two arrays of keys and values must have correct index relations (i.e., first pair on keys[0] and values[0] and so on)!
      *
      * @method
      * @param {string[]} keys
@@ -406,7 +406,7 @@ class CommonMemoryCache extends ConnectionObserver {
     }
 
     /**
-     * Used to check if the provided value is member of the specified set.
+     * Used to check if the provided value is a member of the specified set.
      *
      * @method
      * @param {string} setName
@@ -663,4 +663,4 @@ class CommonMemoryCache extends ConnectionObserver {
 }
 
 const instance = new CommonMemoryCache();
-module.exports = Object.freeze( instance );
+module.exports.instance = Object.freeze( instance );